+++ NEWS.md
@@ -1,4 +1,8 @@
 # ranon (development version)
 
-* Add function `k_anonymity()`
 * Breaking: Refactor function name from `k_anonymisation` to `k_anonymity`
+* Function `k_anonymity()` now properly enforces argument restrictions
+
+# 0.1.0
+
+* Add function `k_anonymity()`